Move channel code to write stage
